requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.10.1
django-filter==21.1
django-heroku==0.0.0
djangorestframework==3.13.1
dropbox==11.25.0
h11==0.12.0
idna==3.3
importlib-metadata==4.10.0
lxml==4.7.1
Markdown==3.3.6
outcome==1.1.0
ply==3.11
pycparser==2.21
pyOpenSSL==21.0.0
pytz==2021.3
requests==2.27.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
stone==3.2.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
whitenoise==5.3.0
wsproto==1.0.0
zipp==3.7.0
psycopg2
gunicorn
